# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Pedidos",
        "description": "API para gestionar pedidos y ventas, incluyendo la creación de pedidos, consulta de productos y entrega.",
        "contact": {
            "email": "soporte@tusistema.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "api/administradores/pedidos/crear": {
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Crear un nuevo pedido",
                "operationId": "36ce0633efd79c802454842be4fdb324",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Pedido"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/administradores/pedidos/listar": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obtener todos los pedidos",
                "operationId": "5da7b608787c2a6ec1cc8500372bbfd8",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pedido"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/administradores/pedidos/buscar/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obtener un pedido por su ID",
                "operationId": "5571860ccb8d4018febb609ae168726d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedido"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/administradores/pedidos/entregar/{id}": {
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Marcar un pedido como entregado",
                "operationId": "258992385b60c308a117637fd67caeac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user": {
                                        "description": "Usuario que realiza la entrega",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido entregado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se entregó el pedido y se generó la venta exitosamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Pedido no encontrado o no en estado PENDIENTE"
                    },
                    "500": {
                        "description": "Error al generar la venta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Pedido": {
                "required": [
                    "id",
                    "estado",
                    "clientes_id",
                    "administradores_id",
                    "cobrado",
                    "direccion",
                    "ventas_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID del pedido",
                        "type": "integer"
                    },
                    "estado": {
                        "description": "Estado del pedido",
                        "type": "string",
                        "enum": [
                            "PENDIENTE",
                            "ENTREGADO"
                        ]
                    },
                    "fecha_creacion": {
                        "description": "Fecha de creación del pedido",
                        "type": "string",
                        "format": "date-time"
                    },
                    "fecha_actualizacion": {
                        "description": "Fecha de actualización del pedido",
                        "type": "string",
                        "format": "date-time"
                    },
                    "clientes_id": {
                        "description": "ID del cliente",
                        "type": "integer"
                    },
                    "administradores_id": {
                        "description": "ID del administrador",
                        "type": "integer"
                    },
                    "cobrado": {
                        "description": "Monto cobrado",
                        "type": "integer"
                    },
                    "direccion": {
                        "description": "Dirección de entrega",
                        "type": "string"
                    },
                    "fecha_entrega": {
                        "description": "Fecha de entrega",
                        "type": "string",
                        "format": "date-time"
                    },
                    "ventas_id": {
                        "description": "ID de la venta asociada al pedido",
                        "type": "integer"
                    },
                    "productos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Producto"
                        }
                    }
                },
                "type": "object"
            },
            "Producto": {
                "required": [
                    "id",
                    "estado",
                    "fecha_creacion",
                    "nombre",
                    "precio_venta",
                    "precio_base",
                    "medida"
                ],
                "properties": {
                    "id": {
                        "description": "ID del producto",
                        "type": "integer",
                        "example": 101
                    },
                    "estado": {
                        "description": "Estado del producto",
                        "type": "string",
                        "enum": [
                            "ACTIVO",
                            "INACTIVO"
                        ],
                        "example": "ACTIVO"
                    },
                    "fecha_creacion": {
                        "description": "Fecha de creación del producto",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-27T10:00:00Z"
                    },
                    "fecha_actualizacion": {
                        "description": "Fecha de actualización del producto",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-27T12:00:00Z"
                    },
                    "nombre": {
                        "description": "Nombre del producto",
                        "type": "string",
                        "example": "Producto A"
                    },
                    "descripcion": {
                        "description": "Descripción del producto",
                        "type": "string",
                        "example": "Este es un producto de ejemplo."
                    },
                    "precio_venta": {
                        "description": "Precio de venta del producto",
                        "type": "number",
                        "format": "float",
                        "example": 20.5
                    },
                    "precio_base": {
                        "description": "Precio base del producto",
                        "type": "number",
                        "format": "float",
                        "example": 15
                    },
                    "medida": {
                        "description": "Unidad de medida del producto",
                        "type": "string",
                        "enum": [
                            "unidades",
                            "metros",
                            "kilos",
                            "litros"
                        ],
                        "example": "kilos"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utiliza un token de autorización Bearer para autenticar las solicitudes.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}